+++ src/commands/index.ts
@@ -1,7 +1,6 @@
 export * from './Article';
 export * from './Backers';
 export * from './Cache';
-export * from './Chatbot';
 export * from './Content';
 export * from './Dashboard';
 export * from './Diagnostics';